WLP: allow unit debug dialog to increase over max_hitpoints
I find it way more useful this way. The maximum allowed values for
hitpoints, experience, moves and attacks_left is now
a factor using factor * max_value.
Perhaps a slider is not the best widget though, maybe simply put
a box there to enter a numerical value. With current and max values
printed into the label.