AtS: fix ranged attack animation glitches affecting the Goliath unit type
Turns out the TORPEDO_BLAST macro's [if]/[else] sequence interacts with
the similar inline sequence that's used for controlling other aspects of
the unit animation. The only solution for now is expanding and embedding
the contents of TORPEDO_BLAST into the animation proper so there's now a
unique big [if]/[else] sequence in the animation.

Along the way, I optimized some aspects to make the torpedo blast halo
animation shorter and easier to maintain.