AtSM: After the Storm Music Resources add-on
This add-on is intended to become a dependency for AtS 0.8.x and
later, once After_the_Storm/music/ is dropped.

It contains music used for the After the Storm campaign, provided
in a separate package in order to allow for faster and more frequent
releases and reduced upload/download times.

Installation of this add-on is optional for users with limited
download rates or quotas, or users who don't play with music or
sounds enabled.

There's currently no WML, and there will probably never be; AtS is
supposed to take care of that area by itself. The music files for
now are copied from After_the_Storm/music/.