AtS: use Project_Y_After_the_Storm as namespace for persistent variables
This used to be Project_Y.After_the_Storm, using a subnamespace as
recommended by the wiki. However, this seems to cause segmentation
faults on Mac OS X (vultraz) and memory corruption on Linux (me) because
of a problem with an implicitly-created, empty [variables] node in the
persistent store WML.

Yes, I know. bugs.wesnoth.org, etc. Humbug.