AtS E1S11.2: fix spurious "retrieving member of non-existent WML container" warning on victory
The warning in question:

> warning engine: variable_info: retrieving member of non-existent WML container, rtw_end_gamestate.player_stats

This is actually caused by storing data on a WML container that doesn't
yet exist that is a child of a WML container that doesn't yet exist
either. Normally, Wesnoth does not complain about storing attributes on
a WML container that doesn't yet exist, but storing a whole WML node
triggers the warning above, even though it actually works correctly in
practice.

The workaround is simple enough:

> {VARIABLE rtw_end_gamestate.non_existent_wml_container_warning_workaround yes}

This forces the parent container to be created (which doesn't trigger a
warning), so the game doesn't complain about its child container's
creation afterwards.

The child container was being created by a [store_side] action, now
preceded by the workaround oneliner.

I am not sure if this is really a bug with the mainline engine's
handling of vconfig stuff or not.