NX: E1 storytext edits.
Note that the Great Contintent is such "of Irdya" the first, not second,
time it is mentioned.

"[Wesnoth's] kingdom" -> "[Wesnoth's] domain" (Wesnoth *is* a kingdom,
it doesn't *have* one.)

Add a comma to a sentence in need of one.

Avoid repetition of "relative calm" and diminish repetition of "across".

Major edits to second part.

Changed second and third parts to present tense; I think this sounds
better, but feel free to revert.

Misc. minor FIs.