TEG: simplify scenario 4
Remove lots of unneccessary variables.
Add scenario status setters.
Scenario ends whenever the first enemy leader dies. If we paid, we only
need to defeat the bandit leader and get to choose the next scenario.
If we didn't, we need to defeat one of them and get to the next scenario
accordingly.