AtS: fix faerie units from old saved games on preload
The event handler searches for all units with race=fairy on the gamemap
and recall lists and updates them to use race=faerie and update several
image paths that aren't automatically updated by the game engine on
load, including the portrait path (for units using the baseframe as
their portrait) and image paths added or replaced by [modifications].

In order to make this work, we need to keep a bare-bones definition of
the old race name around, since the game will otherwise the race for
units with unknown race with the empty string.

This code does not fix units that are stored in variables when loading a
game. I am not sure handling those is worth the additional trouble.