IftU: refactor Althurin's stats definition macro to allow easier inclusion of a canrecruit=yes version in scenario 8; also rename a file
Since the implementation of hero ellipses into the IS_HERO macro in
1.11.x, toggling IS_HERO on and off is no longer just matter of messing
with unit overlays, so we might as well make things easier here by
applying IS_HERO later on instead.

I also renamed a file and committed it in this same commit by accident.
I'm too lazy to use git rebase to amend this now.