{RC_LOAD_ERA} is now lua based and allows selecting any default type era.
Adds two eras to replace what was removed, they are Random Campaign Default & Random Campaign Age of Heroes.(Only difference from default & aoh eras is for the Dwarvish Runesmith line for Knalgan Alliance leader).

Also, flags for default factions are removed, due to the fact that the opponent can be determined by watching side switch at top of screen. The mechanism to use flags is still in place, as adding flags can be done in Custom Campaign, or in an era's factions.

Support for AoH 4 scenario campaign to come later.